+++ packages/core/src/di/interface/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export * from './injection_token';
+export * from './injector';
+export * from './defs';
+export * from './provider';
