+++ packages/core/src/di/r3_injector.ts
@@ -11,10 +11,9 @@ import {OnDestroy} from '../metadata/lifecycle_hooks';
 import {stringify} from '../util/stringify';
 
 import {resolveForwardRef} from './forward_ref';
-import {InjectionToken} from './injection_token';
 import {INJECTOR, Injector, NullInjector, THROW_IF_NOT_FOUND, USE_VALUE} from './injector';
-import {InjectFlags, inject, injectArgs, setCurrentInjector} from './injector_compatibility';
-import {InjectableDef, InjectableType, InjectorType, InjectorTypeWithProviders, getInjectableDef, getInjectorDef} from './interface/defs';
+import {inject, injectArgs, setCurrentInjector} from './injector_compatibility';
+import {InjectFlags, InjectableDef, InjectableType, InjectionToken, InjectorType, InjectorTypeWithProviders, getInjectableDef, getInjectorDef} from './interface';
 import {ClassProvider, ConstructorProvider, ExistingProvider, FactoryProvider, Provider, StaticClassProvider, StaticProvider, TypeProvider, ValueProvider} from './interface/provider';
 import {APP_ROOT} from './scope';
