+++ packages/core/src/render3/interfaces/injector.ts
@@ -6,9 +6,9 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {InjectionToken} from '../../di/injection_token';
-import {InjectFlags} from '../../di/injector_compatibility';
+import {InjectFlags, InjectionToken} from '../../di/interface';
 import {Type} from '../../interface/type';
+
 import {TElementNode} from './node';
 import {LView, TData} from './view';
